moved registry cli things to registry cli
