+++ crates/wash-lib/src/cli/mod.rs
@@ -41,6 +41,7 @@ pub mod get;
 pub mod inspect;
 pub mod link;
 pub mod output;
+pub mod registry;
 pub mod spy;
 pub mod start;
 pub mod stop;
